sqp-ann/requirements.txt
accelerate==0.16.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
asttokens==3.0.0
audioread==3.0.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
comm==0.2.3
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
dotenv==0.9.9
exceptiongroup==1.3.0
executing==2.2.0
expelliarmus==1.1.12
filelock==3.18.0
flatdict==4.0.1
fonttools==4.59.1
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.45
h5py==3.14.0
hf-xet==1.1.7
huggingface-hub==0.34.4
hydra-core==1.3.2
idna==3.10
importRosbag==1.0.4
ipykernel==6.30.1
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.9
lazy_loader==0.4
librosa==0.11.0
lightning-utilities==0.15.2
llvmlite==0.40.1
logging==0.4.9.6
MarkupSafe==3.0.2
matplotlib==3.10.5
matplotlib-inline==0.1.7
mpmath==1.3.0
msgpack==1.1.1
nest-asyncio==1.6.0
networkx==3.4.2
neurobench==2.0.0
nir==1.0.6
nirtorch==2.0.5
numba==0.57.1
numpy==1.24.4
omegaconf==2.3.0
packaging==25.0
pandas==2.3.1
parso==0.8.4
pbr==7.0.0
pexpect==4.9.0
pillow==11.3.0
platformdirs==4.3.8
pooch==1.8.2
prompt_toolkit==3.0.51
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.0.1
requests==2.32.4
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.15.3
sentry-sdk==2.34.1
six==1.17.0
smmap==5.0.2
snntorch==0.9.1
soundfile==0.13.1
soxr==0.5.0.post1
stack-data==0.6.3
sympy==1.14.0
threadpoolctl==3.6.0
tonic==1.6.0
torch==2.8.0
torchaudio
torchinfo
torchmetrics
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
triton==3.4.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
wandb==0.21.1
wcwidth==0.2.13
